feat: refactor chart components
