--- components/charts/HorizontalBarChart.vue
@@ -1,35 +1,0 @@
-<script setup lang="ts">
-import {
-  Chart as ChartJS,
-  BarElement,
-  Tooltip,
-  Legend,
-  CategoryScale,
-  LinearScale,
-} from "chart.js";
-import { Bar } from "vue-chartjs";
-import { useHorizontalBar } from "~/utils/chartConfig";
-
-ChartJS.register(BarElement, Tooltip, Legend, CategoryScale, LinearScale);
-
-const props = defineProps<{
-  title: string;
-  labels: string[];
-  data: number[];
-}>();
-
-const chartConfig = useHorizontalBar(props.labels, props.data);
-</script>
-
-<template>
-  <UCard class="h-full">
-    <UText tag="h1" size="large" weight="medium" class="mt-0">{{
-      title
-    }}</UText>
-    <Bar
-      :data="chartConfig.data"
-      :options="chartConfig.options"
-      class="h-full max-h-96"
-    />
-  </UCard>
-</template>
