+++ components/charts/doughnut-chart.vue
