+++ layouts/default.vue
@@ -1,9 +1,9 @@
 <template>
   <div>
-    <theHeader />
+    <TheHeader />
     <div class="min-h-screen flex flex-col flex-grow p-5 pt-20 md:pt-24">
       <slot />
     </div>
-    <theFooter />
+    <TheFooter />
   </div>
 </template>
